Introduction of param min.alt.ind
This param sets a threshold which dictates if a marker should be
included in the analysis based on the minimum number of individuals
carrying at least a single alt allele according the the current model.
It is means as an alternative to max.geno.freq that is based on raw
count and not frequency.